--- nmea-lib/src/main/java/com/capitanperegrina/nmea/api/parser/NMEAParser.java
@@ -1,7 +1,0 @@
-package com.capitanperegrina.nmea.api.parser;
-
-import com.capitanperegrina.nmea.api.model.beans.GPSInformation;
-
-public interface NMEAParser {
-    public GPSInformation parse(String line);
-}
