--- nmea-lib/src/main/java/com/capitanperegrina/nmea/api/parser/sentenceparser/NMEASentenceParser.java
@@ -1,8 +1,0 @@
-package com.capitanperegrina.nmea.api.parser.sentenceparser;
-
-import com.capitanperegrina.nmea.api.model.beans.GPSInformation;
-
-public interface NMEASentenceParser {
-
-	GPSInformation parse(String [] tokens);
-}
