+++ nmea-lib/src/main/java/com/capitanperegrina/nmea/impl/sentence/parsers/listeners/RMCListener.java
@@ -1,0 +1,14 @@
+package com.capitanperegrina.nmea.impl.sentence.parsers.listeners;
+
+import net.sf.marineapi.nmea.event.AbstractSentenceListener;
+import net.sf.marineapi.nmea.sentence.RMCSentence;
+import net.sf.marineapi.nmea.util.Date;
+import net.sf.marineapi.nmea.util.Position;
+import net.sf.marineapi.nmea.util.Time;
+
+public class RMCListener extends AbstractSentenceListener<RMCSentence> {
+    public void sentenceRead(RMCSentence rmc) {
+        Position pos = rmc.getPosition();
+        System.out.println("[" + rmc.getTime().toISO8601() + "] " + pos.getDatum().toString() + " - " + pos.toString());
+    }
+}
